requirements.txt
opencv-python==4.6.0.66
Pillow==9.2.0
face_recognition==1.3.0
flask==2.2.2
